Move settings into feature.settings, screenshots into feature.screenshots, filter into feafure.filter
