+++ app/src/main/java/ru/d3rvich/jetgames/di/FeatureModule.kt
@@ -8,8 +8,8 @@ import org.koin.core.annotation.Module
     "ru.d3rvich.feature.browse",
     "ru.d3rvich.detail",
     "ru.d3rvich.favorites",
-    "ru.d3rvich.filter",
+    "ru.d3rvich.feature.filter",
     "ru.d3rvich.home",
-    "ru.d3rvich.settings"
+    "ru.d3rvich.feature.settings"
 )
 object FeatureModule
