+++ feature/filter/src/main/java/ru/d3rvich/feature/filter/views/FilterAppBar.kt
@@ -1,0 +1,67 @@
+package ru.d3rvich.feature.filter.views
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import ru.d3rvich.core.ui.theme.JetGamesTheme
+import ru.d3rvich.filter.R
+import ru.d3rvich.common.R as uiR
+
+/**
+ * Created by Ilya Deryabin at 02.04.2024
+ */
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun FilterAppBar(
+    isResetButtonVisible: Boolean,
+    onBackClicked: () -> Unit,
+    onResetClicked: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    TopAppBar(
+        modifier = modifier,
+        title = { Text(text = stringResource(R.string.filter_label)) },
+        navigationIcon = {
+            IconButton(onClick = onBackClicked) {
+                Icon(
+                    painter = painterResource(uiR.drawable.arrow_back_24px),
+                    contentDescription = stringResource(R.string.navigate_back)
+                )
+            }
+        },
+        actions = {
+            AnimatedVisibility(
+                visible = isResetButtonVisible,
+                enter = fadeIn(),
+                exit = fadeOut()
+            ) {
+                TextButton(onClick = onResetClicked) {
+                    Text(text = stringResource(R.string.reset))
+                }
+            }
+        }
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun FilterAppBarPreview() {
+    JetGamesTheme {
+        FilterAppBar(
+            isResetButtonVisible = true,
+            onBackClicked = {},
+            onResetClicked = {}
+        )
+    }
+}
