+++ feature/settings/src/main/java/ru/d3rvich/feature/settings/SettingsViewModel.kt
@@ -1,0 +1,67 @@
+package ru.d3rvich.feature.settings
+
+import android.os.Build
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.launch
+import org.koin.core.annotation.KoinViewModel
+import ru.d3rvich.core.domain.model.ColorModeType
+import ru.d3rvich.core.domain.model.ThemeType
+import ru.d3rvich.core.domain.repositories.UserPreferencesRepository
+import ru.d3rvich.core.ui.base.BaseViewModel
+import ru.d3rvich.core.ui.base.UiAction
+import ru.d3rvich.core.ui.base.UiEvent
+import ru.d3rvich.core.ui.base.UiState
+
+@KoinViewModel
+internal class SettingsViewModel(private val userPreferencesRepository: UserPreferencesRepository) :
+    BaseViewModel<SettingsUiState, SettingsUiEvent, UiAction>() {
+
+    override fun createInitialState(): SettingsUiState = SettingsUiState.Loading
+
+    override fun obtainEvent(event: SettingsUiEvent) {
+        viewModelScope.launch {
+            when (event) {
+                is SettingsUiEvent.UpdateColorMode -> {
+                    userPreferencesRepository.setCurrentColorMode(event.colorModeType)
+                }
+
+                is SettingsUiEvent.UpdateThemeType -> {
+                    userPreferencesRepository.setCurrentTheme(event.themeType)
+                }
+            }
+        }
+    }
+
+    private val isDynamicColorSupported by lazy {
+        Build.VERSION.SDK_INT >= Build.VERSION_CODES.S
+    }
+
+    init {
+        viewModelScope.launch {
+            userPreferencesRepository.getUserPreferences().collect { data ->
+                setState(
+                    SettingsUiState.Settings(
+                        themeType = data.theme,
+                        colorModeType = data.colorMode,
+                        inDynamicColorSupported = isDynamicColorSupported
+                    )
+                )
+            }
+        }
+    }
+}
+
+internal sealed interface SettingsUiState : UiState {
+    data object Loading : SettingsUiState
+    class Settings(
+        val themeType: ThemeType,
+        val colorModeType: ColorModeType,
+        val inDynamicColorSupported: Boolean
+    ) : SettingsUiState
+}
+
+internal sealed interface SettingsUiEvent : UiEvent {
+    class UpdateThemeType(val themeType: ThemeType) : SettingsUiEvent
+
+    class UpdateColorMode(val colorModeType: ColorModeType) : SettingsUiEvent
+}
